Coins: MOVEONLY: Move CCoinsViewCache and CCoinsModifier to coinscache.o
