+++ src/Makefile.am
@@ -85,6 +85,7 @@ BITCOIN_CORE_H = \
   clientversion.h \
   coincontrol.h \
   coins.h \
+  coinscache.h \
   compat.h \
   compressor.h \
   consensus/params.h \
@@ -243,6 +244,7 @@ libbitcoin_common_a_SOURCES = \
   base58.cpp \
   chainparams.cpp \
   coins.cpp \
+  coinscache.cpp \
   compressor.cpp \
   primitives/block.cpp \
   primitives/transaction.cpp \
