+++ src/bitcoin-tx.cpp
@@ -4,18 +4,19 @@
 
 #include "base58.h"
 #include "clientversion.h"
+#include "coins.h"
+#include "coinscache.h"
+#include "core_io.h"
+#include "keystore.h"
 #include "primitives/block.h" // for MAX_BLOCK_SIZE
 #include "primitives/transaction.h"
-#include "core_io.h"
-#include "coins.h"
-#include "keystore.h"
 #include "script/script.h"
 #include "script/sign.h"
 #include "ui_interface.h" // for _(...)
 #include "univalue/univalue.h"
 #include "util.h"
-#include "utilstrencodings.h"
 #include "utilmoneystr.h"
+#include "utilstrencodings.h"
 
 #include <stdio.h>
