+++ src/coins.h
@@ -375,86 +375,4 @@ public:
     const CTxOut& GetOutputFor(const CTxIn& input) const;
 };
 
-class CCoinsViewCache;
-
-/** 
- * A reference to a mutable cache entry. Encapsulating it allows us to run
- *  cleanup code after the modification is finished, and keeping track of
- *  concurrent modifications. 
- */
-class CCoinsModifier
-{
-private:
-    CCoinsViewCache& cache;
-    CCoinsMap::iterator it;
-    CCoinsModifier(CCoinsViewCache& cache_, CCoinsMap::iterator it_);
-
-public:
-    CCoins* operator->() { return &it->second.coins; }
-    CCoins& operator*() { return it->second.coins; }
-    ~CCoinsModifier();
-    friend class CCoinsViewCache;
-};
-
-/** CCoinsView that adds a memory cache for transactions to another CCoinsView */
-class CCoinsViewCache : public CCoinsViewEfficient
-{
-protected:
-    /* Whether this cache has an active modifier. */
-    bool hasModifier;
-
-    /**
-     * Make mutable so that we can "fill the cache" even from Get-methods
-     * declared as "const".  
-     */
-    mutable uint256 hashBlock;
-    mutable CCoinsMap cacheCoins;
-
-public:
-    CCoinsViewCache(CCoinsView *baseIn);
-    virtual ~CCoinsViewCache();
-
-    // Standard CCoinsView methods
-    bool GetCoins(const uint256 &txid, CCoins &coins) const;
-    bool HaveCoins(const uint256 &txid) const;
-    uint256 GetBestBlock() const;
-    void SetBestBlock(const uint256 &hashBlock);
-    bool BatchWrite(CCoinsMap &mapCoins, const uint256 &hashBlock);
-
-    /**
-     * Return a pointer to CCoins in the cache, or NULL if not found. This is
-     * more efficient than GetCoins. Modifications to other cache entries are
-     * allowed while accessing the returned pointer.
-     */
-    virtual const CCoins* AccessCoins(const uint256 &txid) const;
-
-    /**
-     * Return a modifiable reference to a CCoins. If no entry with the given
-     * txid exists, a new one is created. Simultaneous modifications are not
-     * allowed.
-     */
-    CCoinsModifier ModifyCoins(const uint256 &txid);
-
-    /**
-     * Push the modifications applied to this cache to its base.
-     * Failure to call this method before destruction will cause the changes to be forgotten.
-     * If false is returned, the state of this cache (and its backing view) will be undefined.
-     */
-    bool Flush();
-
-    //! Calculate the size of the cache (in number of transactions)
-    unsigned int GetCacheSize() const;
-
-    friend class CCoinsModifier;
-
-private:
-    CCoinsMap::iterator FetchCoins(const uint256 &txid);
-    CCoinsMap::const_iterator FetchCoins(const uint256 &txid) const;
-
-    /**
-     * By making the copy constructor private, we prevent accidentally using it when one intends to create a cache on top of a base cache.
-     */
-    CCoinsViewCache(const CCoinsViewCache &);
-};
-
 #endif // BITCOIN_COINS_H
