+++ src/rpcblockchain.cpp
@@ -4,6 +4,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "checkpoints.h"
+#include "coinscache.h"
 #include "main.h"
 #include "rpcserver.h"
 #include "sync.h"
