+++ src/rpcrawtransaction.cpp
@@ -4,12 +4,14 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "base58.h"
-#include "primitives/transaction.h"
+#include "coins.h"
+#include "coinscache.h"
 #include "core_io.h"
 #include "init.h"
 #include "keystore.h"
 #include "main.h"
 #include "net.h"
+#include "primitives/transaction.h"
 #include "rpcserver.h"
 #include "script/script.h"
 #include "script/sign.h"
