+++ src/test/coins_tests.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "coins.h"
+#include "coinscache.h"
 #include "random.h"
 #include "uint256.h"
 #include "test/test_bitcoin.h"
