+++ src/test/script_P2SH_tests.cpp
@@ -2,6 +2,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "coins.h"
+#include "coinscache.h"
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
