+++ src/test/transaction_tests.cpp
@@ -7,12 +7,14 @@
 #include "test/test_bitcoin.h"
 
 #include "clientversion.h"
+#include "coins.h"
+#include "coinscache.h"
+#include "core_io.h"
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
 #include "script/script.h"
 #include "script/script_error.h"
-#include "core_io.h"
 
 #include <map>
 #include <string>
